chore: factor out screen data override and version code
