+++ lib/screens/solari_screen_data.ex
@@ -1,39 +1,7 @@
 defmodule Screens.SolariScreenData do
   @moduledoc false
 
-  alias Screens.LogScreenData
-
-  def by_screen_id_with_override_and_version(screen_id, client_version, is_screen) do
-    if Screens.Override.State.disabled?(String.to_integer(screen_id)) do
-      LogScreenData.log_api_response(screen_id, client_version, is_screen, %{
-        force_reload: false,
-        success: false
-      })
-    else
-      LogScreenData.log_api_response(
-        screen_id,
-        client_version,
-        is_screen,
-        by_screen_id_with_version(
-          screen_id,
-          client_version,
-          is_screen
-        )
-      )
-    end
-  end
-
-  defp by_screen_id_with_version(screen_id, client_version, is_screen) do
-    api_version = Application.get_env(:screens, :api_version)
-
-    if api_version == client_version do
-      by_screen_id(screen_id, is_screen)
-    else
-      %{force_reload: true}
-    end
-  end
-
-  defp by_screen_id(screen_id, _is_screen) do
+  def by_screen_id(screen_id, _is_screen) do
     %{station_name: station_name, sections: sections} =
       :screens
       |> Application.get_env(:screen_data)
