+++ lib/screens_web/controllers/screen_api_controller.ex
@@ -3,14 +3,6 @@ defmodule ScreensWeb.ScreenApiController do
   require Logger
 
   def show(conn, %{"id" => screen_id, "version" => version}) do
-    config_data =
-      :screens
-      |> Application.get_env(:screen_data)
-      |> Map.get(screen_id)
-
-    app_id = Map.get(config_data, :app_id)
-    stop_id = Map.get(config_data, :stop_id)
-
     is_screen = ScreensWeb.UserAgent.is_screen_conn?(conn)
 
     _ =
@@ -19,38 +11,7 @@ defmodule ScreensWeb.ScreenApiController do
       end
 
     data =
-      case app_id do
-        "bus_eink" ->
-          Screens.BusScreenData.by_stop_id_with_override_and_version(
-            stop_id,
-            screen_id,
-            version,
-            is_screen
-          )
-
-        "gl_eink_single" ->
-          Screens.GLScreenData.by_stop_id_with_override_and_version(
-            stop_id,
-            screen_id,
-            version,
-            is_screen
-          )
-
-        "gl_eink_double" ->
-          Screens.GLScreenData.by_stop_id_with_override_and_version(
-            stop_id,
-            screen_id,
-            version,
-            is_screen
-          )
-
-        "solari" ->
-          Screens.SolariScreenData.by_screen_id_with_override_and_version(
-            screen_id,
-            version,
-            is_screen
-          )
-      end
+      Screens.ScreenData.by_screen_id_with_override_and_version(screen_id, version, is_screen)
 
     json(conn, data)
   end
